# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estude Loterias API - Estatísticas e ferramentas para loterias da Caixa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lotofacil/resultados:
    get:
      operationId: getLotofacilResultados
      tags: [lotofacil]
      summary: Histórico de resultados da Lotofácil
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: ano
          in: query
          schema:
            type: ["integer", "null"]
        - name: ordem
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Lista paginada de resultados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadosPaginadosLotofacil"
  /lotofacil/resultados/{concurso}:
    get:
      operationId: getLotofacilResultadoConcurso
      tags: [lotofacil]
      summary: Resultado de um concurso específico da Lotofácil
      parameters:
        - name: concurso
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resultado do concurso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoLotofacil"
        "404":
          description: Concurso não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /lotofacil/simulador:
    post:
      operationId: simularLotofacil
      tags: [lotofacil]
      summary: Simula uma aposta em todos os concursos anteriores da Lotofácil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimuladorInputLotofacil"
      responses:
        "200":
          description: Resultado da simulação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulacaoResultado"
  /lotofacil/gerador:
    post:
      operationId: gerarJogoLotofacil
      tags: [lotofacil]
      summary: Gera um jogo aleatório de Lotofácil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeradorInputLotofacil"
      responses:
        "200":
          description: Jogo(s) gerado(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JogoGerado"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FaixaPremio:
      type: object
      properties:
        faixa:
          type: integer
        descricao:
          type: string
        ganhadores:
          type: integer
        valorPremio:
          type: number
      required:
        - faixa
        - descricao
        - ganhadores
        - valorPremio
...
    SimulacaoResultado:
      type: object
      required: [resumo, concursos, totalPremio, totalConcursos]
      properties:
        resumo:
          type: array
          items:
            type: object
            required: [acertos, contagem]
            properties:
              acertos:
                type: integer
              contagem:
                type: integer
        resumoTrevos:
          type: array
          items:
            type: object
            required: [acertos, contagem]
            properties:
              acertos:
                type: integer
              contagem:
                type: integer
        faixas:
          type: array
          items:
            type: object
            required: [label, contagem, probabilidade]
            properties:
              label:
                type: string
              contagem:
                type: integer
              probabilidade:
                type: string
        concursos:
          type: array
          items:
            type: object
            required: [concurso, data, dezenas, acertos, premioGanho]
            properties:
              concurso:
                type: integer
              data:
                type: string
              dezenas:
                type: array
                items:
                  type: string
              trevos:
                type: array
                items:
                  type: string
              acertos:
                type: integer
              acertosTrevos:
                type: integer
              premioGanho:
                type: number
        totalPremio:
          type: number
        totalConcursos:
          type: integer
...
    JogoGerado:
      type: object
      properties:
        jogos:
          type: array
          items:
            type: array
            items:
              type: integer
        custo:
          type: number
      required:
        - jogos
        - custo
    ResultadoLotofacil:
      type: object
      properties:
        concurso:
          type: integer
        data:
          type: string
        dezenas:
          type: array
          items:
            type: string
        dezenasOrdem:
          type: array
          nullable: true
          items:
            type: string
        premios:
          type: array
          items:
            $ref: "#/components/schemas/FaixaPremio"
        acumulado:
          type: boolean
        valorAcumulado:
          type: ["number", "null"]
        dataProximoConcurso:
          type: ["string", "null"]
        valorEstimadoProximoConcurso:
          type: ["number", "null"]
        arrecadacaoTotal:
          type: ["number", "null"]
        valorAcumuladoConcurso_0:
          type: ["number", "null"]
        valorAcumuladoConcursoEspecial:
          type: ["number", "null"]
      required:
        - concurso
        - data
        - dezenas
        - premios
        - acumulado
    ResultadosPaginadosLotofacil:
      type: object
      properties:
        total:
          type: integer
        pagina:
          type: integer
        limite:
          type: integer
        totalPaginas:
          type: integer
        resultados:
          type: array
          items:
            $ref: "#/components/schemas/ResultadoLotofacil"
      required:
        - total
        - pagina
        - limite
        - totalPaginas
        - resultados
...
    SimuladorInputLotofacil:
      type: object
      required: [dezenas]
      properties:
        dezenas:
          type: array
          items:
            type: string
          minItems: 15
          maxItems: 20
        filtro:
          type: string
          enum: [todos, premiados, quinze, quatorze, treze, doze, onze]
          default: premiados
    GeradorInputLotofacil:
      type: object
      properties:
        quantidadeJogos:
          type: integer
          minimum: 1
          maximum: 20
          default: 1
        quantidadeDezenas:
          type: integer
          minimum: 15
          maximum: 20
          default: 15
      required:
        - quantidadeJogos
        - quantidadeDezenas
...